+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed [here](https://github.com/sensu-plugins/community/blob/master/HOW_WE_CHANGELOG.md)
 
 # [Unreleased]
+
+# [0.5.2]
 ## Added
 - test that event data exists and is valid json (@barryorourke)
 - make testing plugins a lot easier (@barryorourke)
@@ -87,7 +89,8 @@ This CHANGELOG follows the format listed [here](https://github.com/sensu-plugins
 ## [0.1.0] 2014-01-06
 - Initial release (@zsprackett)
 
-[Unreleased]: https://github.com/sensu-plugins/sensu-plugin-python/compare/0.5.1...HEAD
+[Unreleased]: https://github.com/sensu-plugins/sensu-plugin-python/compare/0.5.2...HEAD
+[0.5.1]: https://github.com/sensu-plugins/sensu-plugin-python/compare/0.5.1...0.5.2
 [0.5.1]: https://github.com/sensu-plugins/sensu-plugin-python/compare/0.5.0...0.5.1
 [0.5.0]: https://github.com/sensu-plugins/sensu-plugin-python/compare/0.4.7...0.5.0
 [0.4.7]: https://github.com/sensu-plugins/sensu-plugin-python/compare/0.4.6...0.4.7